A number of updates primarily relating to handling the insertion of post into the database, handling of embeded content that is inserted into the database, and previewing embeded content before it is inserted.
Some of these changes also have to do with the workflow of what to do when content does or does not work out.

Added a number of new sources as well which content can be pulled from.

Also created a new database entry stored with every post that logs the service origin of each highlight.