Removed the welcome file. Don't need that shit hanging around.
Added the ability to view a single post on a a single page. This page also supports the ability to delete post.

Added the ability to delete post.

Moved some routes around so they can only be accessed by logged in users. This is for adding new post, and removing old ones.

Updated the feed listings to have links to the eventual comment sections that'll be powered with Disquis. This also added a 'Share' button which provides a link to the origin contents perma links, as well as a link to share the unique post from this site.